refactor: split sandbox backend roots
